+++ Cargo.toml
@@ -14,6 +14,7 @@ include = ["src/**/*", "Cargo.toml"]
 atty = "0.2.14"
 clap = "2.33.3"
 httparse = "1.4.1"
+nom = "7.0.0"
 serde_json = "1.0.66"
 serde_yaml = "0.8.17"
 toml = "0.5.8"
